pyproject.toml
[project]
name = "chatbot-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit==2.2.1",
    "langchain==0.3.19",
    "langchain-community==0.3.18",
    "langchain-google-genai==2.0.11",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chatbot_workshop"]
